Wirepas/migration_script/requirements.txt
websocket-client==1.0.0
protobuf==3.17.1
pyyaml==5.4.1